Fix call to R_InitLightTables()
Called at startup, also called when changing r_fov CVAR.

+++ src/d_main.c
@@ -458,7 +458,7 @@ void D_Display(void)
 
         // Figure out how far into the current tic we're in as a fixed_t
         if (vid_capfps != TICRATE)
-            fractionaltic = ((I_GetTimeMS() * TICRATE) % 1000) * FRACUNIT / 1000;
+            fractionaltic = (((int64_t)I_GetTimeMS() * TICRATE) % 1000) * FRACUNIT / 1000;
 
         return;
     }
